Split metadata tag logic by responsibility
