+++ src/gui_app/sample_browser_view/hit_target_tests.rs
@@ -1,4 +1,3 @@
-
 use super::*;
 
 fn message_from(output: Option<WidgetOutput>) -> SampleFileHitMessage {
